requirements.txt
pyTelegramBotAPI==4.26.0
requests
python-dotenv
pycryptodome
mss
pillow
termcolor
opencv-python